# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Koine Gateway API
  version: 1.0.0
  description: HTTP gateway service for Claude Code CLI. Provides REST endpoints for text generation, structured object generation, and streaming responses.
  license:
    name: AGPL-3.0-only OR LicenseRef-Commercial
    url: https://github.com/pattern-zones-co/koine/blob/main/LICENSE
  contact:
    name: Pattern Zones Co
    url: https://github.com/pattern-zones-co/koine
servers:
  - url: http://localhost:3100
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "API key for authentication. Generate with: openssl rand -hex 32"
...
paths:
...
  /generate-text:
    post:
      summary: Generate text response
      description: Generates plain text response from Claude CLI.
      tags:
        - Generation
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                system:
                  type: string
                prompt:
                  type: string
                sessionId:
                  type: string
                model:
                  type: string
                maxTokens:
                  type: integer
                  exclusiveMinimum: 0
              required:
                - prompt
      responses:
        "200":
          description: Successful text generation
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
                  usage:
                    type: object
                    properties:
                      inputTokens:
                        type: integer
                        minimum: 0
                      outputTokens:
                        type: integer
                        minimum: 0
                      totalTokens:
                        type: integer
                        minimum: 0
                    required:
                      - inputTokens
                      - outputTokens
                      - totalTokens
                  sessionId:
                    type: string
                required:
                  - text
                  - usage
                  - sessionId
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  code:
                    type: string
                    enum: *a4
                  rawText:
                    type: string
                required:
                  - error
                  - code
        "401":
          description: Missing authorization header
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "403":
          description: Invalid API key
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "500":
          description: Internal server error or CLI error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  code:
                    type: string
                    enum: *a4
                  rawText:
                    type: string
                required:
                  - error
                  - code
  /generate-object:
    post:
      summary: Generate structured JSON response
      description: Generates structured JSON response from Claude CLI. The schema is passed in the request to instruct Claude to output valid JSON.
      tags:
        - Generation
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                system:
                  type: string
                prompt:
                  type: string
                sessionId:
                  type: string
                model:
                  type: string
                schema:
                  type: object
                  properties:
                    type:
                      type: string
                      enum: *a1
                    properties:
                      type: object
                      additionalProperties: {}
                    items: {}
                    $ref:
                      type: string
                    allOf:
                      type: array
                      items: {}
                    anyOf:
                      type: array
                      items: {}
                    oneOf:
                      type: array
                      items: {}
                    enum:
                      type: array
                      items: {}
                    const: {}
                maxTokens:
                  type: integer
                  exclusiveMinimum: 0
              required:
                - prompt
                - schema
      responses:
        "200":
          description: Successful object generation
          content:
            application/json:
              schema:
                type: object
                properties:
                  object: {}
                  rawText:
                    type: string
                  usage:
                    type: object
                    properties:
                      inputTokens:
                        type: integer
                        minimum: 0
                      outputTokens:
                        type: integer
                        minimum: 0
                      totalTokens:
                        type: integer
                        minimum: 0
                    required:
                      - inputTokens
                      - outputTokens
                      - totalTokens
                  sessionId:
                    type: string
                required:
                  - rawText
                  - usage
                  - sessionId
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  code:
                    type: string
                    enum: *a4
                  rawText:
                    type: string
                required:
                  - error
                  - code
        "401":
          description: Missing authorization header
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "403":
          description: Invalid API key
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "500":
          description: Internal server error or CLI error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  code:
                    type: string
                    enum: *a4
                  rawText:
                    type: string
                required:
                  - error
                  - code
...